Add Matlab experiment for image colour segmentation
We convert the image from RGB colour space to CIE L*a*b values
Using the ab from the Lab colour space to segment the image
The a axis represent the green-red component and the b axis represent
the blue-yellow component.
We have good result with all the apples in the tree are detected, even
the one with occlusion.
There is still noise or false detection, but I think we can reduce the
noise by shape detection.